requirements.txt
boto3==1.9.223
botocore==1.12.223
Django==2.1
django-colorful==1.3
django-extensions==2.1.7
django-storages==1.7.1
django-taggit==1.1.0
django-timetable==0.1.7
docutils==0.15.2
jmespath==0.9.4
Pillow==6.0.0
psycopg2-binary==2.8.3
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
whitenoise==4.1.2
